Latin morphology API documentation

Morphological properties

Tabulae.LMPTense — Type

Tense of finite, infinitive or participial form of verb.

Tabulae.LMPMood — Type

Mood of finite verb form.

Tabulae.LMPVoice — Type

Voice of finite, infinitive or participial form of verb.

Tabulae.LMPPerson — Type

Person of finite verb form.

Tabulae.LMPNumber — Type

Number of substantive or of finite verb form.

Tabulae.LMPGender — Type

Gender of substantive form.

Tabulae.LMPCase — Type

Case of substantive form.

Tabulae.LMPDegree — Type

Degree of adjective or adverb.

Tabulae.LMPUninflectedType — Type

Type of uninflected form ("part of speech").

Constructor functions:

Tabulae.lmpTense — Function

Find tense of a form.

lmpTense(lmf)

Create LMPTense from descriptive string.

lmpTense(s)

Create LMPTense from integer code.

lmpTense(code)

Find tense of a form.

lmpTense(verb)

Tabulae.lmpMood — Function

Find mood of a form.

Create LMPMood from descriptive string.

lmpMood(s)

Create LMPMood from integer code.

lmpMood(code)

Find mood of a form.

lmpMood(verb)

Tabulae.lmpVoice — Function

Find voice of a form.

Create LMPVoice from descriptive string.

lmpVoice(s)

Create LMPVoice from integer code.

lmpVoice(code)

Find voice of a form.

lmpVoice(verb)

Tabulae.lmpPerson — Function

Find person of a form.

Create LMPPerson from descriptive string.

lmpPerson(s)

Create LMPPerson from integer code.

lmpPerson(code)

Find person of a form.

lmpPerson(verb)

Tabulae.lmpNumber — Function

Find number of a form.

Create LMPPerson from descriptive string.

lmpNumber(s)

Create LMPPerson from integer code.

lmpNumber(code)

Find number of a LMFNoun.

lmpNumber(noun)

Find number of a form.

lmpNumber(verb)

Tabulae.lmpGender — Function

Find gender of a form.

Create LMPGender from descriptive string.

lmpGender(s)

Create LMPGender from integer code.

lmpGender(code)

Find gender of a LMFNoun.

lmpGender(noun)

Gender for a TabulaeNounStem

lmpGender(n)

Tabulae.lmpCase — Function

Find case of a form.

Create LMPGender from descriptive string.

lmpCase(s)

Create LMPGender from integer code.

lmpCase(code)

Find case of a LMFNoun.

lmpCase(noun)

Tabulae.lmpDegree — Function

Find degree of a form.

Create LMPPerson from descriptive string.

lmpDegree(s)

Create LMPPerson from integer code.

lmpDegree(code)

Tabulae.lmpUninflectedType — Function

Find part of speech of an uninflected form.

Create LMPGender from descriptive string.

lmpUninflectedType(s)

Create LMPGender from integer code.

lmpUninflectedType(code)

Morphological forms

Tabulae.LatinMorphologicalForm — Type

Abstract type of a morphological form.

Tabulae.LMFNoun — Type

Nouns have gender, case and number.

Tabulae.LMFFiniteVerb — Type

Finite verbs have person, number, tense, mood and voice.

Tabulae.LatinMorphologicalProperty — Type

Abstract type of a morphological form in Tabulae.

Tabulae.code — Function

Integer code for morphological property value.

code(lmp)

Integer code for tense value.

code(tense)

Integer code for a LMPMood value.

code(mood)

Integer code for a LMPVoice value.

code(voice)

Integer code for a LMPPerson value.

code(person)

Integer code for a LMPPerson value.

code(num)

Integer code for a LMPGender value.

code(gender)

Integer code for a LMPCase value.

code(cs)

Integer code for a LMPPerson value.

code(degree)

Integer code for a LMPUninflectedType value.

code(uninfl)

Sequence of digits encoding form noun

code(noun)

Sequence of digits encoding form verb

code(verb)