Functions for working with FormUrns

Tabulae.personcode — Function

Extract person code from a FormUrn.

personcode(f)

Tabulae.personlabel — Function

Look up label for person code from a FormUrn.

personlabel(f)

Tabulae.numbercode — Function

Extract number code from a FormUrn.

numbercode(f)

Tabulae.numberlabel — Function

Look up label for number code from a FormUrn.

numberlabel(f)

Tabulae.tensecode — Function

Extract tense code from a FormUrn.

tensecode(f)

Tabulae.tenselabel — Function

Look up label for tense code from a FormUrn.

tenselabel(f)

Tabulae.moodcode — Function

Extract mood code from a FormUrn.

moodcode(f)

Tabulae.moodlabel — Function

Look up label for mood code from a FormUrn.

moodlabel(f)

Tabulae.voicecode — Function

Extract voice code from a FormUrn.

voicecode(f)

Tabulae.voicelabel — Function

Look up label for voice code from a FormUrn.

voicelabel(f)

Tabulae.gendercode — Function

Extract gender code from a FormUrn.

gendercode(f)

Tabulae.genderlabel — Function

Look up label for gender code from a FormUrn.

genderlabel(f)

Tabulae.casecode — Function

Extract case code from a FormUrn.

casecode(f)

Tabulae.caselabel — Function

Look up label for case code from a FormUrn.

caselabel(f)

Tabulae.degreecode — Function

Extract case code from a FormUrn.

degreecode(f)

Tabulae.degreelabel — Function

Look up label for degree code from a FormUrn.

degreelabel(f)

Tabulae.uninflectedcode — Function

Extract case code from a FormUrn.

uninflectedcode(f)

Tabulae.uninflectedlabel — Function

Look up label for uninflected code from a FormUrn.

uninflectedlabel(f)